+++ ice/load.go
@@ -35,11 +35,11 @@ func load(data *segment.Data) (*Segment, error) {
 	rv := &Segment{
 		data:           data.Slice(0, data.Len()-footerLen),
 		footer:         footer,
-		fieldsMap:      make(map[string]uint16),
-		fieldDvReaders: make(map[uint16]*docValueReader),
-		fieldFSTs:      make(map[uint16]*vellum.FST),
-		fieldDocs:      make(map[uint16]uint64),
-		fieldFreqs:     make(map[uint16]uint64),
+		fieldsMap:      make(map[string]uint8),
+		fieldDvReaders: make(map[uint8]*docValueReader),
+		fieldFSTs:      make(map[uint8]*vellum.FST),
+		fieldDocs:      make(map[uint8]uint64),
+		fieldFreqs:     make(map[uint8]uint64),
 	}
 
 	// FIXME temporarily map to existing footer fields
@@ -120,9 +120,9 @@ func (s *Segment) loadFields() error {
 
 		name := string(nameData)
 		s.fieldsInv = append(s.fieldsInv, name)
-		s.fieldsMap[name] = uint16(fieldID + 1)
-		s.fieldDocs[uint16(fieldID)] = fieldDocVal
-		s.fieldFreqs[uint16(fieldID)] = fieldFreqVal
+		s.fieldsMap[name] = uint8(fieldID + 1)
+		s.fieldDocs[uint8(fieldID)] = fieldDocVal
+		s.fieldFreqs[uint8(fieldID)] = fieldFreqVal
 
 		fieldID++
 	}
